Fix: Se deben enviar las coordenadas como double

+++ app/src/main/java/bdisfer1410/controldepresencia/api/clock/send/ClockSendRequest.java
@@ -1,19 +1,19 @@
 package bdisfer1410.controldepresencia.api.clock.send;
 
 public class ClockSendRequest {
-    private final float latitude;
-    private final float longitude;
+    private final double latitude;
+    private final double longitude;
 
-    public ClockSendRequest(float latitude, float longitude) {
+    public ClockSendRequest(double latitude, double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 }
